apps/neuralsense/requirements.txt
paho-mqtt>=2.0.0
numpy
scipy